Move TEdgeConfigDto into a single file again
